--- example-injects/wargames
@@ -1,11 +1,0 @@
-
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
-  <head>
-    <meta http-equiv="content-type" content="application/xhtml+xml; charset=iso-8859-1" />
-    <title>Inject PWN</title>
-  </head>
-  <body>
-        <h2>Your browser has been hijacked by Airpwn-NG.</h2>
-        <img src="https://pbs.twimg.com/media/B9wl1fdCcAAbSwx.jpg:large" />
-  </body>
