+++ payloads/ballsy
@@ -1,0 +1,12 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml">
+  <head>
+    <meta http-equiv="content-type" content="application/xhtml+xml; charset=iso-8859-1" />
+    <title>Inject PWN</title>
+    <meta content="hacking, hackers, pwn, packet-injection" name="keywords" />
+    <meta content="Packet Injection 101" name="description" />
+  </head>
+  <body>
+    <iframe src="http://www.domain.com"></iframe>
+  </body>
+</html>
