+++ payloads/beef-hooks/pad
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Untitled Page</title>
+    <script type="text/javascript">
+        window.onbeforeunload = confirmBrowseAway;
+        function confirmBrowseAway()
+        {
+          return "blah"
+        }
+        
+  /*      var LoopCounter = 0;
+        function doLoop()
+        {
+            if (LoopCounter < 4)
+            {
+                LoopCounter++;
+                alert(LoopCounter);
+                MyTimout = setTimeout(doLoop, 2000);
+                //MyTimout = setTimeout("doLoop()", 2000);
+            }
+
+            else
+            {
+                clearTimeout(MyTimout);
+            }
+        }*/
+    </script>
+</head>
+<!-- <body onload="doLoop()"> -->
+</body>
+</html>
