+++ payloads/beef-hooks/pad2
@@ -1,0 +1,28 @@
+<html>
+<head>
+<script>
+var workIsDone = false;
+
+window.onbeforeunload = confirmBrowseAway;
+
+function confirmBrowseAway()
+{
+  if (!workIsDone) {
+    return "Are you sure you want to do that? If you leave this page " +
+    "now, your work will NOT be saved!";
+  }
+}
+
+function trigger()
+{
+  // The user has finished their work, so we don't
+  // need a warning anymore!
+  workIsDone = true;
+}
+</script>
+</head>
+<body>
+<h1>Flip the trigger</h1>
+<input type="submit" value="Flip the trigger" onclick="trigger()" />
+</body>
+</html>
